refactor(website): user search component
