+++ apps/website/src/components/dialogs/gift-subscription-dialog.svelte
@@ -5,12 +5,11 @@
 	import Dialog, { type DialogMode } from "./dialog.svelte";
 	import { t } from "svelte-i18n";
 	import Spinner from "../spinner.svelte";
-	import TextInput from "../input/text-input.svelte";
 	import { CaretLeft, User as UserIcon } from "phosphor-svelte";
 	import { type SubscriptionProductVariant, type User } from "$/gql/graphql";
 	import UserName from "../user-name.svelte";
-	import ChannelPreview from "../channel-preview.svelte";
 	import { variantUnit } from "$/lib/utils";
+	import UserSearch from "../user-search.svelte";
 
 	interface Props {
 		mode: DialogMode;
@@ -20,128 +19,6 @@
 	let { mode = $bindable("hidden"), variant }: Props = $props();
 
 	let recipient = $state<User>();
-	let query = $state("");
-
-	let timeout: NodeJS.Timeout | number | undefined; // not reactive
-
-	async function search(query: string): Promise<User[]> {
-		if (!query) {
-			return [];
-		}
-
-		// Small timeout to prevent spamming requests when user is typing
-		return new Promise((resolve, reject) => {
-			if (timeout) {
-				clearTimeout(timeout);
-			}
-
-			timeout = setTimeout(async () => {
-				const res = await gqlClient()
-					.query(
-						graphql(`
-							query SearchUser($query: String!) {
-								users {
-									search(query: $query, page: 1, perPage: 10) {
-										items {
-											id
-											mainConnection {
-												platformDisplayName
-												platformAvatarUrl
-											}
-											roleIds
-											style {
-												activeProfilePicture {
-													images {
-														url
-														mime
-														size
-														width
-														height
-														scale
-														frameCount
-													}
-												}
-												activePaint {
-													id
-													name
-													data {
-														layers {
-															id
-															ty {
-																__typename
-																... on PaintLayerTypeSingleColor {
-																	color {
-																		hex
-																	}
-																}
-																... on PaintLayerTypeLinearGradient {
-																	angle
-																	repeating
-																	stops {
-																		at
-																		color {
-																			hex
-																		}
-																	}
-																}
-																... on PaintLayerTypeRadialGradient {
-																	repeating
-																	stops {
-																		at
-																		color {
-																			hex
-																		}
-																	}
-																	shape
-																}
-																... on PaintLayerTypeImage {
-																	images {
-																		url
-																		mime
-																		size
-																		scale
-																		width
-																		height
-																		frameCount
-																	}
-																}
-															}
-															opacity
-														}
-														shadows {
-															color {
-																hex
-															}
-															offsetX
-															offsetY
-															blur
-														}
-													}
-												}
-											}
-											highestRoleColor {
-												hex
-											}
-										}
-									}
-								}
-							}
-						`),
-						{ query },
-					)
-					.toPromise();
-
-				if (res.error || !res.data) {
-					reject();
-					return;
-				}
-
-				resolve(res.data.users.search.items as User[]);
-			}, 200);
-		});
-	}
-
-	let results = $derived(search(query));
 
 	let giftLoading = $state(false);
 
@@ -184,32 +61,18 @@
 				Gift 1 {variantUnit(variant)} to <UserName user={recipient} />.
 			</p>
 		{:else}
-			<TextInput type="text" placeholder="Search User" bind:value={query}>
+			<UserSearch
+				placeholder="Search User"
+				onResultClick={(e, user) => {
+					e.preventDefault();
+					recipient = user;
+				}}
+			>
 				{#snippet icon()}
-					{#await results}
-						<Spinner />
-					{:then _}
-						<UserIcon />
-					{/await}
+					<UserIcon />
 				{/snippet}
 				<h2>Select recipient</h2>
-			</TextInput>
-			{#await results then results}
-				{#if results && results.length > 0}
-					<div class="results">
-						{#each results as result}
-							<ChannelPreview
-								user={result}
-								size={2}
-								onclick={(e) => {
-									e.preventDefault();
-									recipient = result;
-								}}
-							/>
-						{/each}
-					</div>
-				{/if}
-			{/await}
+			</UserSearch>
 		{/if}
 
 		<div class="buttons">
@@ -272,30 +135,5 @@
 		display: flex;
 		align-items: center;
 		gap: 0.5rem;
-	}
-
-	.results {
-		background-color: var(--bg-light);
-
-		border: 1px solid var(--border-active);
-		border-radius: 0.5rem;
-
-		display: flex;
-		overflow: hidden;
-
-		flex-direction: column;
-
-		& > :global(.button) {
-			animation: expand-down 0.2s forwards;
-		}
-	}
-
-	@keyframes expand-down {
-		from {
-			height: 2rem;
-		}
-		to {
-			height: 2.75rem;
-		}
 	}
 </style>
